Routes reorganized and some fixes: - Responsive design - Plus button position - Create folders for general structure
