+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <meta charset="UTF-8">
-        <link rel="icon" href="/logo-white.svg">
+        <link rel="icon" href="/img/logo-white.svg">
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css">
         <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
