+++ src/components/avatar/stepper/EmailStep.vue
