+++ src/components/avatar/stepper/IdStep.vue
