+++ src/components/avatar/stepper/NameStep.vue
