+++ src/components/avatar/stepper/Success.vue
