+++ src/components/avatar/stepper/VehicleStep.vue
