+++ src/components/citizen/stepper/packages/PackageWeight.vue
